api/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cpu 
torchvision==0.15.0+cpu
matplotlib
requests
uvicorn==0.20.0
gunicorn==19.9.0
fastapi
pydantic
numpy
Pillow==9.3.0
tensorflow-cpu==2.11.0
pathlib==1.0.1
opencv-python
pytz